+++ _axpy/src/utils.c
@@ -9,7 +9,7 @@
 #include "utils.h"
 
 
-void capture_ref_result(double *y, double* y_ref, int n)
+void capture_ref_result(float *y, float* y_ref, int n)
 {
    int i;
    //printf ("\nReference result: ");
@@ -21,13 +21,13 @@ void capture_ref_result(double *y, double* y_ref, int n)
 }
 
 
-void test_result(double *y, double *y_ref, long nrows)
+void test_result(float *y, float *y_ref, long nrows)
 {
    long row;
    int nerrs=0;
    /* Compute with the result to keep the compiler for marking the code as dead */
    for (row=0; row<nrows; row++) {
-      double error = y[row] - y_ref[row];
+      float error = y[row] - y_ref[row];
       if (fabs(error) > 0.0000001)  {
          printf("y[%ld]=%.16f != y_ref[%ld]=%.16f  INCORRECT RESULT !!!! \n ", row, y[row], row, y_ref[row]);
          nerrs++;
