+++ _axpy/src/utils.h
@@ -1,3 +1,3 @@
 
-void capture_ref_result(double *y, double* y_ref, int n);
-void test_result(double *y, double *y_ref, long nrows);
+void capture_ref_result(float *y, float* y_ref, int n);
+void test_result(float *y, float *y_ref, long nrows);
